[Fix] previous Button from first part to last system
When loading a second file with less system than the previous one, jumping with the "previous" button from the first part to the last system failed as it had stored the old size of systems and navigated to a system number which didn't existed. (Likely also caused issues when the number of systems was different and jumped to a system different from the last one but the one with the index number = size of the previous file).

Change-Id: Ie58c38891eb894c29fd504d8ca231bc0bfdf768d